first experiment with complete flow of singleUserVault done. With tests

+++ contracts/test/unit/SingleUserVault.OpenPosition.t.sol
@@ -2,17 +2,18 @@
 pragma solidity ^0.8.24;
 
 import "forge-std/Test.sol";
-import {stdStorage, StdStorage} from "forge-std/StdStorage.sol";
+import { stdStorage, StdStorage } from "forge-std/StdStorage.sol";
 
-import {SingleUserVault} from "../../src/core/SingleUserVault.sol";
-import {MockNFPM} from "../mocks/MockNFPM.sol";
-import {MockPool} from "../mocks/MockPool.sol";
-import {VaultErrors} from "../../src/errors/VaultErrors.sol";
+import { SingleUserVault } from "../../src/core/SingleUserVault.sol";
+import { MockNFPM } from "../mocks/MockNFPM.sol";
+import { MockPool } from "../mocks/MockPool.sol";
+import { VaultErrors } from "../../src/errors/VaultErrors.sol";
 
 /// @title OpenInitialPosition unit tests
 /// @notice Covers guard paths that do not require NFPM integration success.
 contract SingleUserVault_OpenPosition_Test is Test {
     using stdStorage for StdStorage;
+
     StdStorage private _stdstore;
 
     SingleUserVault vault;
@@ -45,10 +46,7 @@ contract SingleUserVault_OpenPosition_Test is Test {
 
     function test_Revert_WhenPositionAlreadyOpened() public {
         // Define positionTokenId = 1 usando stdstore (slot-safe)
-        _stdstore
-            .target(address(vault))
-            .sig(vault.positionTokenId.selector)
-            .checked_write(uint256(1));
+        _stdstore.target(address(vault)).sig(vault.positionTokenId.selector).checked_write(uint(1));
 
         vm.prank(owner);
         vm.expectRevert(VaultErrors.PositionAlreadyOpened.selector);
